webadmin/requirements.txt
https://github.com/darklow/django-suit/tarball/v2
django==2.2.2
pillow
django-ckeditor
django-cleanup
